Adjusted API + Session=>Summed + BacktraceSession=>BacktraceMetrics
