+++ Runtime/BacktraceClient.cs
@@ -47,17 +47,17 @@ namespace Backtrace.Unity
             }
         }
 
-        private IBacktraceSession _session;
+        private BacktraceMetrics _metrics;
         /// <summary>
-        /// Backtrace session instance
+        /// Backtrace metrics instance
         /// </summary>
-        public IBacktraceSession Session
+        public IBacktraceMetrics Metrics
         {
             get
             {
-                if (_session == null && Configuration != null && Configuration.EnableEventAggregationSupport)
+                if (_metrics == null && Configuration != null && Configuration.EnableEventAggregationSupport)
                 {
-                    _session = new BacktraceSession(
+                    _metrics = new BacktraceMetrics(
                         AttributeProvider,
                         Configuration.GetEventAggregationUrl(),
                         Configuration.GetEventAggregationIntervalTimerInMs())
@@ -65,11 +65,7 @@ namespace Backtrace.Unity
                         IgnoreSslValidation = Configuration.IgnoreSslValidation
                     };
                 }
-                return _session;
-            }
-            internal set
-            {
-                _session = value;
+                return _metrics;
             }
         }
 
@@ -495,35 +491,36 @@ namespace Backtrace.Unity
                 nativeCrashUplaoder.SetBacktraceApi(BacktraceApi);
                 StartCoroutine(nativeCrashUplaoder.SendUnhandledGameCrashesOnGameStartup());
             }
-            if (Configuration.EnableEventAggregationSupport)
+            if (Configuration.EnableEventAggregationSupport && Metrics != null)
             {
-                Session.SendStartupEvent();
+                _metrics.SendStartupEvent();
             }
         }
 
-        public void EnableSessionAggregationSupport()
+        public void EnableMetrics()
         {
             if (!Configuration.EnableEventAggregationSupport)
             {
                 Debug.LogWarning("Event aggregation configuration was disabled. Enabling it manually via API");
             }
-            EnableSessionAggregationSupport(Configuration.GetEventAggregationUrl(), Configuration.TimeIntervalInMin);
+            EnableMetrics(Configuration.GetEventAggregationUrl(), Configuration.GetEventAggregationIntervalTimerInMs());
         }
-        public void EnableSessionAggregationSupport(string submissionUrl, long timeIntervalInMs)
+        public void EnableMetrics(string submissionUrl, long timeIntervalInMs = BacktraceMetrics.DefaultTimeIntervalInMs, string uniqueEventName = BacktraceMetrics.DefaultUniqueEventName)
         {
-            if (Session != null)
+            if (_metrics != null)
             {
-                Debug.LogWarning("Backtrace session aggregation support is enabled. Please use BacktraceClient.Session.");
+                Debug.LogWarning("Backtrace metrics support is enabled. Please use BacktraceClient.Metrics.");
                 return;
             }
-            Session = new BacktraceSession(
+            _metrics = new BacktraceMetrics(
                 attributeProvider: AttributeProvider,
                 uploadUrl: submissionUrl,
                 timeIntervalInMs: timeIntervalInMs)
             {
+                StartupUniqueEventName = uniqueEventName,
                 IgnoreSslValidation = Configuration.IgnoreSslValidation
             };
-            Session.SendStartupEvent();
+            _metrics.SendStartupEvent();
         }
 
         private void OnApplicationQuit()
@@ -542,7 +539,7 @@ namespace Backtrace.Unity
         private void LateUpdate()
         {
             _nativeClient?.UpdateClientTime(Time.unscaledTime);
-            Session?.Tick(Time.unscaledTime);
+            _metrics?.Tick(Time.unscaledTime);
 
             if (BackgroundExceptions.Count == 0)
             {
