+++ Runtime/Interfaces/IBacktraceClient.cs
@@ -10,9 +10,9 @@ namespace Backtrace.Unity.Interfaces
     public interface IBacktraceClient
     {
         /// <summary>
-        /// Backtrace session instance
+        /// Backtrace Metrics instance
         /// </summary>
-        IBacktraceSession Session { get; }
+        IBacktraceMetrics Metrics { get; }
         /// <summary>
         /// Send a new report to a Backtrace API
         /// </summary>
@@ -50,11 +50,11 @@ namespace Backtrace.Unity.Interfaces
         /// <summary>
         /// Enable event aggregation support.
         /// </summary>
-        void EnableSessionAggregationSupport();
+        void EnableMetrics();
 
         /// <summary>
         /// Enable event aggregation support.
         /// </summary>
-        void EnableSessionAggregationSupport(string submissionUrl, long timeIntervalInMs);
+        void EnableMetrics(string submissionUrl, long timeIntervalInMs, string uniqueEventName);
     }
 }
