+++ Runtime/Model/BacktraceConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using Backtrace.Unity.Common;
+using Backtrace.Unity.Services;
 using Backtrace.Unity.Types;
 using System;
 using System.Collections.Generic;
@@ -188,7 +189,7 @@ namespace Backtrace.Unity.Model
         /// </summary>
         [Range(0, 60)]
         [Tooltip("How often events should be sent to the Backtrace endpoints, in minutes. Zero (0) disables auto send and will require manual periodic sending using the API. For more information, see the README.")]
-        public long TimeIntervalInMin = 30;
+        public long TimeIntervalInMin = BacktraceMetrics.DefaultTimeIntervalInMin;
 
         /// <summary>
         /// Determine if database is enable
