--- Runtime/Model/Session/SessionSubmissionJob.cs
@@ -1,10 +1,0 @@
-﻿namespace Backtrace.Unity.Model.Session
-{
-    internal sealed class SessionSubmissionJob
-    {
-        public double NextInvokeTime { get; set; }
-        public UniqueEvent[] UniqueEvents { get; set; }
-        public SessionEvent[] SessionEvents { get; set; }
-        public uint NumberOfAttemps { get; set; }
-    }
-}
