+++ Runtime/Services/BacktraceMetrics.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: be44b3b4cd99cf7438bd5063c8b6a7fb
+guid: abba691b0fdb5ec419b05dcd3d9c4056
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
