--- Tests/Runtime/Session/BacktraceSessionTests.cs
@@ -1,332 +1,0 @@
-﻿using Backtrace.Unity.Json;
-using Backtrace.Unity.Model.JsonData;
-using Backtrace.Unity.Services;
-using Backtrace.Unity.Tests.Runtime.Session.Mocks;
-using NUnit.Framework;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.TestTools;
-
-namespace Backtrace.Unity.Tests.Runtime.Session
-{
-    public class BacktraceSessionTests
-    {
-        const string SessionEventName = "scene-changed";
-        // existing attribute name in Backtrace
-        const string UniqueAttributeName = "scene.name";
-        private const string _submissionUrl = "https://event-edge.backtrace.io/api/user-aggregation/events?token=TOKEN";
-        private AttributeProvider _attributeProvider = new AttributeProvider();
-
-        [Test]
-        public void BacktraceSession_ShouldTriggerUploadProcessOnSendMethodWithOnlySessionEvent_DataWasSendToTheService()
-        {
-            var jsonString = string.Empty;
-            var submissionUrl = string.Empty;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0);
-            backtraceSession.RequestHandler = new BacktraceHttpClientMock()
-            {
-                OnInvoke = (string url, BacktraceJObject json) =>
-                {
-                    jsonString = json.ToJson();
-                    submissionUrl = url;
-                }
-            };
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.Send();
-
-            Assert.IsNotEmpty(jsonString);
-            Assert.AreEqual(submissionUrl, _submissionUrl);
-            Assert.IsEmpty(backtraceSession.SessionEvents);
-        }
-
-        [Test]
-        public void BacktraceSession_ShouldTriggerUploadProcessOnSendMethodWithOnlyUniqueEvent_DataWasSendToTheService()
-        {
-            var jsonString = string.Empty;
-            var submissionUrl = string.Empty;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0);
-            backtraceSession.RequestHandler = new BacktraceHttpClientMock()
-            {
-                OnInvoke = (string url, BacktraceJObject json) =>
-                {
-                    jsonString = json.ToJson();
-                    submissionUrl = url;
-                }
-            };
-
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-            backtraceSession.Send();
-
-            Assert.IsNotEmpty(jsonString);
-            Assert.AreEqual(submissionUrl, _submissionUrl);
-            Assert.IsEmpty(backtraceSession.SessionEvents);
-        }
-
-
-        [Test]
-        public void BacktraceSession_ShouldntTriggerUploadWhenDataIsNotAvailable_DataWasntSendToBacktrace()
-        {
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0);
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.Send();
-
-            Assert.IsFalse(requestHandler.Called);
-        }
-
-        [UnityTest]
-        public IEnumerator BacktraceSession_ShouldTry3TimesOn503BeforeDroppingEvents_DataWasntSendToBacktrace()
-        {
-            const int expectedNumberOfEventsAfterFailure = 2;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0);
-            var requestHandler = new BacktraceHttpClientMock()
-            {
-                StatusCode = 503
-            };
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-            backtraceSession.Send();
-            for (int i = 0; i < BacktraceSession.MaxNumberOfAttemps; i++)
-            {
-                yield return new WaitForSeconds(1);
-                // immidiately run next update
-                var time = BacktraceSession.MaxTimeBetweenRequests + (BacktraceSession.MaxTimeBetweenRequests * i) + i + 1;
-                backtraceSession.Tick(time);
-            }
-
-            yield return new WaitForSeconds(1);
-            Assert.AreEqual(BacktraceSession.MaxNumberOfAttemps, requestHandler.NumberOfRequests);
-            Assert.AreEqual(backtraceSession.Count(), expectedNumberOfEventsAfterFailure);
-        }
-
-
-        [UnityTest]
-        public IEnumerator BacktraceSession_ShouldTry3TimesOn503AndDropSessionEventsOnMaximumNumberOfEvents_DataWasDeleted()
-        {
-            const int expectedNumberOfEventsAfterFailure = 1;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0);
-            var requestHandler = new BacktraceHttpClientMock()
-            {
-                StatusCode = 503
-            };
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-            backtraceSession.MaximumEvents = expectedNumberOfEventsAfterFailure;
-
-            backtraceSession.Send();
-            for (int i = 0; i < BacktraceSession.MaxNumberOfAttemps; i++)
-            {
-                yield return new WaitForSeconds(1);
-                // immidiately run next update
-                var time = BacktraceSession.MaxTimeBetweenRequests + (BacktraceSession.MaxTimeBetweenRequests * i) + i + 1;
-                backtraceSession.Tick(time);
-            }
-
-            yield return new WaitForSeconds(1);
-            Assert.AreEqual(BacktraceSession.MaxNumberOfAttemps, requestHandler.NumberOfRequests);
-            Assert.AreEqual(backtraceSession.Count(), expectedNumberOfEventsAfterFailure);
-        }
-
-        [Test]
-        public void BacktraceSession_ShouldTryOnlyOnceOnHttpFailure_DataWasntSendToBacktrace()
-        {
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0);
-            var requestHandler = new BacktraceHttpClientMock()
-            {
-                IsHttpError = true
-            };
-            backtraceSession.RequestHandler = requestHandler;
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-            backtraceSession.Send();
-
-            Assert.AreEqual(requestHandler.NumberOfRequests, 1);
-        }
-
-
-        [Test]
-        public void BacktraceSession_ShouldSkipMoreUniqueEventsWhenHitTheLimit_DataWasntSendToBacktrace()
-        {
-            const int maximumNumberOfEvents = 3;
-            const int numberOfTestEvents = 10;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0)
-            {
-                MaximumEvents = maximumNumberOfEvents
-            };
-
-            for (int i = 0; i < numberOfTestEvents; i++)
-            {
-                backtraceSession.AddUniqueEvent($"{UniqueAttributeName} {i}", new Dictionary<string, string>() { { $"{UniqueAttributeName} {i}", "value" } });
-            }
-
-            Assert.AreEqual(maximumNumberOfEvents, backtraceSession.Count());
-        }
-
-        [Test]
-        public void BacktraceSession_ShouldSkipMoreSessionEventsWhenHitTheLimit_DataWasntSendToBacktrace()
-        {
-            const int maximumNumberOfEvents = 3;
-            const int numberOfTestEvents = 10;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 0)
-            {
-                MaximumEvents = maximumNumberOfEvents
-            };
-
-            for (int i = 0; i < numberOfTestEvents; i++)
-            {
-                backtraceSession.AddSessionEvent($"{SessionEventName} {i}");
-            }
-
-            Assert.AreEqual(maximumNumberOfEvents, backtraceSession.Count());
-        }
-
-        [Test]
-        public void BacktraceSession_ShouldTriggerUploadViaTickMethodWhenReachedMaximumNumberOfEvents_DataWasSendToBacktrace()
-        {
-            const int maximumNumberOfEvents = 3;
-            const int defaultTimeIntervalInMs = 10;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, defaultTimeIntervalInMs)
-            {
-                MaximumEvents = maximumNumberOfEvents
-            };
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-            for (int i = 0; i < maximumNumberOfEvents; i++)
-            {
-                backtraceSession.AddSessionEvent($"{SessionEventName} {i}");
-            }
-
-            backtraceSession.Tick(defaultTimeIntervalInMs + 1);
-
-            Assert.AreEqual(0, backtraceSession.Count());
-            Assert.AreEqual(1, requestHandler.NumberOfRequests);
-        }
-
-        [Test]
-        public void BacktraceSession_ShouldntTriggerDownloadViaTickMethodWhenDidntReachMaximumNumberOfEvents_DataWasSendToBacktrace()
-        {
-            const int maximumNumberOfEvents = 3;
-            const int expectedNumberOfEvents = 2;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 1)
-            {
-                MaximumEvents = maximumNumberOfEvents
-            };
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-            for (int i = 0; i < expectedNumberOfEvents; i++)
-            {
-                backtraceSession.AddSessionEvent($"{SessionEventName} {i}");
-            }
-
-            backtraceSession.Tick(0);
-
-            Assert.AreEqual(expectedNumberOfEvents, backtraceSession.Count());
-            Assert.AreEqual(0, requestHandler.NumberOfRequests);
-        }
-
-
-        [Test]
-        public void BacktraceSession_ShouldTriggerUploadAfterTimeIntervalHit_DataWasSendToBacktrace()
-        {
-            const int timeInterval = 10;
-            const int expectedNumberOfEvents = 1;// unique event
-            const int expectedNumberOfRequests = 1; //should combine data together
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, timeInterval);
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-
-            backtraceSession.Tick(timeInterval + 1);
-
-
-            Assert.AreEqual(expectedNumberOfEvents, backtraceSession.Count());
-            Assert.AreEqual(expectedNumberOfRequests, requestHandler.NumberOfRequests);
-        }
-
-        [Test]
-        public void BacktraceSession_ShouldntTriggerDownloadBeforeTimeIntervalHit_DataWasSendToBacktrace()
-        {
-            const int timeInterval = 10;
-            const int numberOfAddedEvents = 2;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, timeInterval);
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-
-            backtraceSession.Tick(timeInterval - 1);
-
-            Assert.AreEqual(numberOfAddedEvents, backtraceSession.Count());
-            Assert.AreEqual(0, requestHandler.NumberOfRequests);
-        }
-
-
-        [Test]
-        public void BacktraceSession_ShouldTriggerUploadAfterTimeIntervalHitAgain_DataWasSendToBacktrace()
-        {
-            const int timeInterval = 10;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, timeInterval);
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-
-            backtraceSession.Tick(timeInterval + 1);
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-
-            backtraceSession.Tick((timeInterval * 2) + 1);
-
-            // we added two unique events - but because we added two the same reports
-            // they should be combined
-            Assert.AreEqual(1, backtraceSession.Count());
-            Assert.AreEqual(2, requestHandler.NumberOfRequests);
-        }
-        [Test]
-        public void BacktraceSession_ShouldntTriggerDownloadAfterTimeIntervalFirstHit_DataWasSendToBacktrace()
-        {
-            const int timeInterval = 10;
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, timeInterval);
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-
-            backtraceSession.Tick(timeInterval + 1);
-
-            backtraceSession.AddSessionEvent(SessionEventName);
-            backtraceSession.AddUniqueEvent(UniqueAttributeName);
-
-            backtraceSession.Tick(timeInterval + 2);
-
-            Assert.AreEqual(2, backtraceSession.Count());
-            Assert.AreEqual(1, requestHandler.NumberOfRequests);
-        }
-
-        [Test]
-        public void BacktraceSessionDefaultEvent_ShouldSendDefaultEventOnTheApplicationStartup_DataWasSendToBacktrace()
-        {
-            var backtraceSession = new BacktraceSession(_attributeProvider, _submissionUrl, 10);
-            var requestHandler = new BacktraceHttpClientMock();
-            backtraceSession.RequestHandler = requestHandler;
-
-            backtraceSession.SendStartupEvent();
-
-            Assert.AreEqual(1, requestHandler.NumberOfRequests);
-        }
-
-    }
-}
